# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi untuk Service API Admin Loker",
        "title": "Admin Loker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7778",
    "paths": {
...
        "/api/v1/level-pekerjaan/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create Level Pekerjaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Level Pekerjaan"
                ],
                "summary": "Create Level Pekerjaan",
                "parameters": [
                    {
                        "description": "Create Level Pekerjaan",
                        "name": "LevelPekerjaan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeLevelPekerjaanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses201Example"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/level-pekerjaan/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Delete Level Pekerjaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Level Pekerjaan"
                ],
                "summary": "Get Delete Level Pekerjaan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Level Pekerjaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/level-pekerjaan/result/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get By ID Level Pekerjaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Level Pekerjaan"
                ],
                "summary": "Get By ID Level Pekerjaan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get By ID Level Pekerjaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/level-pekerjaan/results": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get List Level Pekerjaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Level Pekerjaan"
                ],
                "summary": "Get List Level Pekerjaan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/level-pekerjaan/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Update Level Pekerjaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Level Pekerjaan"
                ],
                "summary": "Get Update Level Pekerjaan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Update Level Pekerjaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Level Pekerjaan",
                        "name": "LevelPekerjaan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeLevelPekerjaanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.SchemeResponses500Example"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemes.SchemeLevelPekerjaanRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "Input with Lowercase",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Staff"
                }
            }
        },
...
        "schemes.SchemeResponses": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status OK"
                }
            }
        },
        "schemes.SchemeResponses201Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 201
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Create Successfully"
                }
            }
        },
        "schemes.SchemeResponses400Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "schemes.SchemeResponses401Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "message": {
                    "type": "string",
                    "example": "Authorization header is required"
                }
            }
        },
        "schemes.SchemeResponses403Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status forbidden"
                }
            }
        },
        "schemes.SchemeResponses404Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status not found"
                }
            }
        },
        "schemes.SchemeResponses409Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 409
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status conflict data"
                }
            }
        },
        "schemes.SchemeResponses500Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status internal error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "User JWT Bearer - Format Input Value : Bearer[ space ][ JWTToken ]",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}